feat(charm): add UA validation and support for attachment of ubuntu-advantage charm
This commit adds the UA validation for the charm. This is accomplished
via a new relation called `ua-attach` which supports the interface
shared by the `ubuntu-advantage` charm. We need the interface as we want
to trigger the deferred installation of the AMS charm as soon as the
machine gets attached to UA.