fix(tests): provide snap as resource to integration tests
This commit adds the capability to supply the ams snap to the
integration tests. The snap will be first downloaded before running the
integration tests. The download requires a secret to download the snap
from the snap store which should be configured in the repository under
github secrets as `SNAPSTORE_TOKEN`.