ci(tests): remove the requirement for snapcraft token to download the snap
This commit removes the requirement for a token to be configured on the
repository to download the snap resource required to test the charm.
This is a pre cursor to completely getting rid of the snap resource
required for the charm as the snap can be directly downloaded from the
snap store.