feat(cos): add cos-agent relation
This commit adds relation for integrating with COS (Canonical
Observability Stack). The relation is formed to the `grafana-agent` charm which
acts as the proxy to the COS bundle deployed and related to the
`grafana-agent` charm through cross-model relations.
This commit only adds support for exporting metrics to prometheus. The
ams logs can be searched through the syslog which is already being
scraped by the `grafana-agent` running on the host.