feat(charm): remove the dependency of snap resource
This commit removes the charm's dependency over the need for a snap
resource. The charm now fetches the `ams` snap from the snap store as it
has been made unlisted. The charm now has two configuration options
added `snap_revision`, `snap_channel` to customize the installation for
the snap. By default the revision for the snap will correspond to the
latest revision in the track available from the `versions` file and the default
risk used is `stable`.